+++ src/kernel/lib/lib.h
