+++ src/kernel/lib/memory.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "lib.h"
 
 void MemoryCopy(void* dest, const void* src, Size size) {
     u8* dest8 = (u8*)dest;
