+++ src/kernel/lib/print.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "lib.h"
 #include "../common/common.h"
 #include "../console/console.h"
