+++ src/kernel/lib/string.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "lib.h"
 
 Size StringLength(const char* str) {
     Size len = 0;
