+++ src/kernel/lib/util.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "lib.h"
 
 void Suspend() {
     while (1) {
